First Commit
A linked list is a dynamic data structure. The number of nodes in a list is not fixed and can grow and shrink on demand. Any application which has to deal with an unknown number of objects will need to use a linked list.
Operations:
Create Node
Insert at the start
Inset at position
Delete the first node
Delete the last node
Delete at position
Display the link list
+++++++++++++++++++++++++++++++++
Hasnain Gakahar